@@ -1,56 +1,45 @@
-# https://github.com/python-poetry/poetry#introduction
+# https://python-poetry.org/docs/
 
 [tool.poetry]
-name = "weatherbot"
-version = "0.0.0"
-description = ""
+name = "weather-smartapp"
+version = "0.1.0"
+description = "weather_smartapp"
 authors = []
 
 
 [tool.poetry.dependencies]
-python = "~3.8"
+python = "~3.9"
 
-botx = "~0.28.0"
-fastapi = "~0.61.2"
-uvicorn = "~0.12.3"
-python-multipart = "~0.0.5"
+pybotx = "~0.32.0"
+pybotx-smartapp-rpc = "~0.3.0"
 
-tortoise-orm = "~0.16.18"
-databases = { version = "~0.4.1", extras = ["postgresql"] }
-alembic = "~1.3.2"
-psycopg2-binary = "~2.8.6"  # for alembic
+fastapi = "~0.70.1"
+uvicorn = "~0.16.0"
 
-loguru = "~0.5.3"
-mako = "~1.1.0"
+aioredis = "~1.3.1"
+loguru = "~0.6.0"
+mako = "~1.1.4"
+pydantic = { version = "~1.8.2", extras = ["dotenv"] }
 
-sentry-sdk = "~0.17.3"
-aiofiles = "^0.6.0"
+alembic = "~1.7.5"
+SQLAlchemy = "~1.4.28"
+asyncpg = "~0.25.0"  # DB backend for application
+psycopg2-binary = "~2.9.2"  # DB backend for alembic (migration tool)
 
 
 [tool.poetry.dev-dependencies]
-black = "~20.8b1"
-isort = "4.3.21"
-autoflake = "~1.4"
-mypy = "~0.782"
-wemake-python-styleguide = "~0.14"
+add-trailing-comma = "2.2.1"
+autoflake = "1.4.0"
+black = "21.12b0"
+isort = "5.10.1"
+mypy = "0.910.0"
+wemake-python-styleguide = "0.16.0"
+click = "8.0.1"
+bandit = "1.7.2"  # https://github.com/PyCQA/bandit/issues/837
 
-docker = "~4.0"
-asgi-lifespan = "~1.0.0"
-
-pytest = "~5.4.1"
-pytest-cov = "~2.8.1"
-pytest-asyncio = "~0.12.0"
-pytest-localserver = "~0.5.0"
-python-dotenv = "~0.13.0"
-pytest-env = "~0.6.2"
-pytest-clarity = "~0.3.0-alpha.0"
-pytest-timeout = "~1.3.4"
-
-molten = "1.0.1"
-redis = "3.5.3"
-tenacity = "6.2.0"
+# TODO: Add pytest
 
 
 [build-system]
-requires = ["poetry>=1.0"]
+requires = ["poetry>=1.1.12"]
 build-backend = "poetry.masonry.api"
