@@ -1,43 +1,59 @@
 """Handlers for default bot commands and system events."""
 
-from botx import Bot, Collector, Message, SendingMessage
+from os import environ
 
+from pybotx import (
+    Bot,
+    BubbleMarkup,
+    ChatCreatedEvent,
+    HandlerCollector,
+    IncomingMessage,
+    SmartAppEvent,
+    StatusRecipient,
+)
+
+from app.bot.smartapp import smartapp
 from app.resources import strings
 
-collector = Collector()
+collector = HandlerCollector()
 
 
-@collector.default(include_in_status=False)
-async def default_handler() -> None:
-    """Run if command not found."""
+@collector.smartapp_event
+async def handle_smartapp_event(event: SmartAppEvent, bot: Bot) -> None:
+    await smartapp.handle_smartapp_event(event, bot)
 
 
 @collector.chat_created
-async def chat_created(message: Message, bot: Bot) -> None:
+async def chat_created_handler(event: ChatCreatedEvent, bot: Bot) -> None:
     """Send a welcome message and the bot functionality in new created chat."""
-    text = strings.CHAT_CREATED_TEMPLATE.format(bot_name=bot.state.bot_name)
-    reply = SendingMessage.from_message(text=text, message=message)
-    reply.add_bubble(command=bot.command_for("help"), label=strings.HELP_LABEL)
-    await bot.send(reply)
+
+    answer_body = strings.CHAT_CREATED_TEMPLATE.format(
+        bot_project_name=strings.BOT_DISPLAY_NAME
+    )
+    bubbles = BubbleMarkup()
+    bubbles.add_button(command="/help", label=strings.HELP_LABEL)
+
+    await bot.answer_message(answer_body, bubbles=bubbles)
 
 
-@collector.handler(
-    command="/help", name="help", description=strings.HELP_COMMAND_DESCRIPTION
-)
-async def show_help(message: Message, bot: Bot) -> None:
-    """Справка по командам."""
-    status = await message.bot.status()
+@collector.command("/help", description="Get available commands")
+async def help_handler(message: IncomingMessage, bot: Bot) -> None:
+    """Show commands list."""
 
-    # For each public command:
-    # * collect full description or
-    # * collect short description like in status or
-    # * skip command without any description
-    commands = []
-    for command in status.result.commands:
-        command_handler = message.bot.handler_for(command.name)
-        description = command_handler.full_description or command_handler.description
-        if description:
-            commands.append((command.body, description))
+    status_recipient = StatusRecipient.from_incoming_message(message)
 
-    text = strings.HELP_COMMAND_MESSAGE_TEMPLATE.format(commands=commands)
-    await bot.answer_message(text, message)
+    status = await bot.get_status(status_recipient)
+    command_map = dict(sorted(status.items()))
+
+    answer_body = "\n".join(
+        f"`{command}` -- {description}" for command, description in command_map.items()
+    )
+
+    await bot.answer_message(answer_body)
+
+
+@collector.command("/_debug:git-commit-sha", visible=False)
+async def git_commit_sha(message: IncomingMessage, bot: Bot) -> None:
+    """Show git commit SHA."""
+
+    await bot.answer_message(environ.get("GIT_COMMIT_SHA", "<undefined>"), message)
