@@ -8,17 +8,18 @@ from sqlalchemy import engine_from_config, pool
 # init config
 sys.path.append(str(pathlib.Path(__file__).resolve().parents[3]))
 
-from app.settings.config import get_app_settings  # isort:skip
+# Imports from `app` should go after `path` patch
+from app.settings import settings  # isort:skip
+from app.db.sqlalchemy import Base, make_url_sync  # isort:skip
 
-config = get_app_settings()
+# Import models to make them visible by alembic
+import app.db.weather.models  # isort:skip
 
+postgres_dsn = make_url_sync(settings.POSTGRES_DSN)
 context_config = context.config
-
 fileConfig(context_config.config_file_name)
-
-target_metadata = None
-
-context_config.set_main_option("sqlalchemy.url", str(config.DATABASE_URL))
+target_metadata = Base.metadata
+context_config.set_main_option("sqlalchemy.url", postgres_dsn)
 
 
 def run_migrations_online() -> None:
