@@ -1,36 +1,35 @@
 """Application with configuration for events, routers and middleware."""
-from botx import UnknownBotError
+
 from fastapi import FastAPI
 
-from app.api.error_handlers import unknown_bot
 from app.api.routers import router
 from app.bot.bot import bot
+from app.db.sqlalchemy import build_db_session_factory
+from app.resources import strings
 from app.services.static_files import StaticFilesCustomHeaders
-from app.settings.config import get_app_settings
-from app.settings.events import shutdown, startup
 
-config = get_app_settings()
+
+async def startup() -> None:
+    # -- Bot --
+    await bot.startup()
+
+    # -- Database --
+    bot.state.db_session_factory = await build_db_session_factory()
+
+
+async def shutdown() -> None:
+    # -- Bot --
+    await bot.shutdown()
 
 
 def get_application() -> FastAPI:
     """Create configured server application instance."""
-    application = FastAPI(title="weatherbot")
-    application.state = bot.state
+
+    application = FastAPI(title=strings.BOT_PROJECT_NAME)
     application.state.bot = bot
 
-    application.add_event_handler(
-        "startup",
-        startup(
-            db_dsn=config.DATABASE_URL,
-            bot_app=bot,
-        ),
-    )
-
-    application.add_event_handler("shutdown", shutdown(bot_app=bot))
-
-    application.add_exception_handler(
-        UnknownBotError, unknown_bot.message_from_unknown_bot_handler
-    )
+    application.add_event_handler("startup", startup)
+    application.add_event_handler("shutdown", shutdown)
 
     application.include_router(router)
 
